+++ lending/RELEASES.md
@@ -399,3 +399,13 @@ Based on:
 - [typescript v0.4.16] lending
 ### Releases
 - [NPM v0.4.16] https://www.npmjs.com/package/ryan-lending/v/0.4.16 - lending
+
+## 2024-10-24 00:58:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.421.1 (2.438.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.4.17] lending
+### Releases
+- [NPM v0.4.17] https://www.npmjs.com/package/ryan-lending/v/0.4.17 - lending